backend/requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
greenlet>=3.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
mcp>=1.25,<2
httpx>=0.27.0

# Authentication dependencies
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin bcrypt version for passlib compatibility
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9

# Development dependencies (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.2.0
aiosqlite>=0.19.0
